test(cli): the LSP transport pins descend from bin_smoke (the cap breathes)
bin_smoke.rs sat at 1499/1500 — the next test would have tripped the
file cap mid-train. The two stdio transport-survival pins (the #282
block-bomb + adversarial positions) move to lsp_transport.rs, their
natural seam (transport robustness, not verb-contract smoke); the two
tiny helpers are duplicated by design (integration files share no
code). 38+2 green, clippy clean, ~136 lines of headroom restored.
